feat: Add Zapier-style universal integration system with AI-powered field mapping
- Add complete integration management system with CRUD operations
- Implement drag-and-drop field mapping canvas with visual transformation builder
- Add AI-powered field mapping suggestions and data transformation recommendations
- Create external API field discovery service with fallback to mock data
- Add comprehensive integration logging and monitoring
- Implement webhook triggers for enquiry and quote data push
- Add external API endpoints for ERP system integration
- Support multiple authentication types (API Key, Bearer Token, Basic Auth, OAuth2)
- Add integration dashboard with real-time statistics and management
- Include transformation preview and validation capabilities
- Add support for bidirectional data sync (push/pull)
- Implement integration status tracking and error handling
- Add mock data generation for demo purposes when external APIs are unreachable

Backend changes:
- New Integration, IntegrationLog entities with full CRUD
- IntegrationController with field discovery and AI suggestions
- FieldDiscoveryService with external API introspection
- IntegrationService for lifecycle management
- ExternalApiController for ERP system access
- Enhanced models with new enums and DTOs
- Updated existing services to trigger integration webhooks

Frontend changes:
- IntegrationDashboard with tabbed interface
- FieldMappingCanvas with drag-and-drop functionality
- TransformationBuilder with visual rule creation
- AI suggestion panels and transformation previews
- Integration management UI with create/edit/delete operations
- Admin-only integration tab in main dashboard

This implements a complete Zapier-style integration platform within the application.

+++ backend/src/main/java/com/procost/api/model/OrderItem.java
@@ -21,7 +21,7 @@ public class OrderItem {
     @Column(columnDefinition = "TEXT")
     private String itemDescription;
     
-    private Integer quantity;
+    private Double quantity;
     private Double unitPrice;
     private Double totalPrice;
     
@@ -41,7 +41,7 @@ public class OrderItem {
         this.order = order;
         this.quoteItem = quoteItem;
         this.itemDescription = quoteItem.getItemDescription();
-        this.quantity = quoteItem.getQuantity();
+        this.quantity = quoteItem.getQuantity().doubleValue();
         this.unitPrice = quoteItem.getUnitPrice();
         this.totalPrice = quoteItem.getTotalPrice();
         this.currency = quoteItem.getCurrency();
@@ -80,11 +80,11 @@ public class OrderItem {
         this.itemDescription = itemDescription;
     }
     
-    public Integer getQuantity() {
+    public Double getQuantity() {
         return quantity;
     }
     
-    public void setQuantity(Integer quantity) {
+    public void setQuantity(Double quantity) {
         this.quantity = quantity;
         updateTotalPrice();
     }
